Attempted to improve Fractal performance
I was going to drill all the way down the recursive tree, create a PShape for the clock hands, then copy it (for the other 1-2 hands and larger for the center) as I stepped back up to create the fractal. Sadly, it looks like you can't copy a PShape for some reason. Maybe if I figured that out I can revisit this method.